Move operator lifecycle under internal
Interfaces changed:

Before:

  import "github.com/valon-technologies/gestalt/server/services/operator"

After:

  import "github.com/valon-technologies/gestalt/server/internal/operator"

Example:

  lc := operator.NewLifecycle()

  cfg, _, err := lc.LoadForExecutionAtPathsWithStatePaths(paths, operator.StatePaths{}, locked)

This keeps the daemon-owned config/operator lifecycle out of production service paths and removes the services/operator package that was only consumed by internal/daemon.

Functional/integration checks:

  go test ./internal/operator -count=1
  go test ./internal/daemon -run 'TestE2EValidateAcceptsCanonicalConfigShapes|TestE2EValidateAcceptsV3TelemetryBuiltins' -count=1
  go list ./...
  golangci-lint run ./internal/operator ./internal/daemon
  git diff --check
