+++ gestaltd/internal/operator/lifecycle_test.go
