+++ gestaltd/internal/operator/localconfig.go
