+++ gestaltd/internal/operator/localconfig_test.go
